Groundwork for web page
In this commit I add the groundwork for displaying the data fetched by
the python scripts in a web page.

It consists of the standard HTML, CSS and JavaScript trio. I decided not
to use any fancy library to keep things simple at the beginning.

During development, you'll need a web server to serve the page and its
resources. You can use python's http.server to do this:

    $ cd html
    $ python3 -m http.server 8000

The page will be available at http://localhost:8000/.

Although no fancy library is used, the JavaScript code uses some modern
features so the page may not work on older browsers.

As the plotting library I chose [C3.js](https://c3js.org/) because I had
some familiarity with it, but using it was harder than I expected.