Don't rely on the ord_d field
The `ord_d` field starts at 0 for `rj` and `all` datasets for the year
of 2018. For these datasets the first date of 20018 is in September, so I believe
there is a bug in the data.

Since this value can be easily computed from the `d` field with luxon,
and this is something that was mentioned in #5, I stopped relying on the
`ord_d` field.